refactor: corrigido os autoimport e a origem do import usando @
