+++ backend/src/modules/user-management/student/domain/entity/user-student.entity.ts
@@ -1,4 +1,4 @@
-import { isGreaterZero, isNumeric } from '@/src/util/validations';
+import { isGreaterZero, isNumeric } from '@/util/validations';
 import UserBase, {
   UserBaseProps,
 } from '../../../@shared/domain/entity/user-base.entity';
